Building A Custom React Component
This lesson covers creating a component, exporting it and importing it
into another component for use.

The components directory has been created.
Within it, a js file called ExpenseItem is our first custom component.
This component has an export command attached to it which allows us to
import it into the App.js file, which is the component responsible for
rendering the application.